Add files via upload
In this procedure we scale the components of a feature vector such that the complete vector has a 
length of 1 or, in other words a norm of 1. Note that this normalisation procedure normalises the 
feature vector, and not the observation vector. So we divide by the norm of the feature vector, 
observation per observation, across the different variables, and not by the norm of the observation 
vector, across observations for the same feature.